backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
sqlparse==0.5.5
tzdata==2025.3
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
python-dotenv
groq
Pillow
